refactor: DTO로 변경
